requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
passlib==1.7.4
protobuf==3.15.6
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
wtf==0.1
WTForms==2.3.3
